Split internal/common/values into input_values, arguments
